Serial 通訊: 1.此版本只是測試兩點校正方法的測試函數尚未導入ADC這邊需要修復ADC功能 2.需要留意slop斜率計算尚在轉至成int型別時候會發生數值太小都已0計算這邊需要放大一定倍率之後再縮放回去才能正常計算

+++ Calibattion_TEST_TO_GUI/Core/Inc/variable.h
@@ -50,6 +50,6 @@ extern uint32_t Data_5V_Max_Addr;
 extern uint32_t Data_12V_Min_Addr;
 extern uint32_t Data_12V_Max_Addr;
 /*********************Slope method**********************/
-extern uint32_t Slope_value;
+extern float Slope_value;
 
 #endif
